# Conflicts:
#	ops_bron_rek.f90
#	ops_depoparexp.f90
#	ops_main.f90
#	ops_plot_uitv.f90
#	ops_print_info.f90
#	ops_reken.f90
#	ops_src_char.f90

+++ ops_reken.f90
@@ -125,8 +125,8 @@ real,      INTENT(IN)                            :: so2bg_rcp                  !
 real,      INTENT(IN)                            :: rhno3_rcp                  ! ratio [HNO3]/[NO3]_total at receptor points, [NO3]_total = [HNO3] + [NO3_aerosol]
 real,      INTENT(IN)                            :: bqrv                       ! source strength of space heating source (rv << "ruimteverwarming" = space heating) [g/s]
 real,      INTENT(IN)                            :: bqtr                       ! source strength of traffic source [g/s]
-INTEGER*4, INTENT(IN)                            :: bx                         ! x-coordinate of source
-INTEGER*4, INTENT(IN)                            :: by                         ! y-coordinate of source
+real,      INTENT(IN)                            :: bx                         ! x-coordinate of source
+real,      INTENT(IN)                            :: by                         ! y-coordinate of source
 real,      INTENT(IN)                            :: bdiam                      ! source diameter [m]; if bdiam < 0 -> circular source, bdiam > 0 -> square sourc
 real,      INTENT(IN)                            :: bsterkte                   ! source strength [g/s]
 real,      INTENT(IN)                            :: bwarmte                    ! heat content of source [MW]
